Added reddit; added debugging option; use tmp file for json

+++ tuiapis.sh
@@ -1,4 +1,20 @@
 #!/bin/bash
+set -e
+
+DEBUG="false"
+
+for i in "$@"
+do
+case $i in
+    --debug)
+    DEBUG="true"
+    shift
+    ;;
+    *)
+          # unknown option
+    ;;
+esac
+done
 
 NAME=$1
 shift
@@ -26,13 +42,25 @@ seturl
 
 seturlparams
 
+if [ $DEBUG == "true" ]; then
+  VERBOSE="-vv"
+else
+  VERBOSE="--silent"
+fi
+
 # fetch data
-curl -o $TMPDIR/tmp_$NAME.txt --silent $URLPARAMS $URL
+curl -o $TMPDIR/tmp_$NAME.txt $VERBOSE $URLPARAMS $URL
+
+if [ ! -f $TMPDIR/tmp_$NAME.txt ]
+then
+  echo 'Could not fetch result.'
+  exit 0
+fi
 
 # build json from tmp file
-# JSON="$("$HELPERDIR/JSON.sh" -b < "$TMPDIR/tmp_$NAME.txt")"
-JSON="$(echo -e $TMPDIR/tmp_$NAME.txt | awk -f $HELPERDIR/JSON.awk)"
+echo -e $TMPDIR/tmp_$NAME.txt | awk -f $HELPERDIR/JSON.awk > $TMPDIR/tmp_$NAME.json
 
 printResults
 
 rm "$TMPDIR/tmp_$NAME.txt"
+rm "$TMPDIR/tmp_$NAME.json"
